Use the :requires_new option for nested transactions
Nested transactions are not really supported at the database level.
Because of this, nesting transaction blocks can cause
the outer block to be aborted which will cause later SQL statements
to fail with the message:

PG::InFailedSqlTransaction: ERROR: current transaction is aborted,
commands ignored until end of transaction block

This is currently the case with the buffered_committer's 1000 change
transaction and the one started in the attempt_change method in the
OneWayReplicator.

To deal with this the PostgreSQLAdapter#transaction method accepts
a `requires_new` option which uses save points to simulate nested
transactions.

This change adds this option to the previously nested transaction
in the OneWayReplicator

This will allow for the insert to fail (for example with a unique
constraint error) and the outer transaction to continue.

More information about this option can be found at:
http://api.rubyonrails.org/classes/ActiveRecord/ConnectionAdapters/DatabaseStatements.html#method-i-transaction